refactor: extract assignGameServerWithRetry to own file
