+++ src/games/plugins/launch-new-game.ts
@@ -4,7 +4,8 @@ import { QueueState } from '../../database/models/queue-state.model'
 import { queue } from '../../queue'
 import { debounce } from 'es-toolkit'
 import { safe } from '../../utils/safe'
-import { launchGame, assignGameServerWithRetry } from '../launch-game'
+import { launchGame } from '../launch-game'
+import { assignGameServerWithRetry } from '../assign-game-server-with-retry'
 import { GameState } from '../../database/models/game.model'
 import { configure, cancelConfigure } from '../rcon/configure'
 import { collections } from '../../database/collections'
